Split out signJWT
