+++ src/resolvers/mutation.ts
@@ -4,7 +4,7 @@ import {
 	authorizeRoleAccess,
 	authorizeOwnership,
 } from '../helpers/authorization.js'
-import {signJWT} from '../helpers/secrets.js'
+import {signJWT} from '../helpers/sign-jwt.js'
 import {Courses, Projects, Users} from '../models.js'
 import type {MutationResolvers as Resolvers} from '../types/generated/schema.js'
 import {Role} from '../types/generated/schema.js'
